refactor(kube): collapse duplicated CR-list helpers into listCachedCRs
flux.listCachedCRs, argocd_projects.listFromCRWatch, the inline block in
ListArgoApplications and karpenter.listKarpenterCR were four copies of
the same cached-CR lookup. Keep one canonical listCachedCRs in crd.go
(returning []*unstructured.Unstructured) and route the others through it.
No behavior change.
